refactor: 연결 플랫폼 관련 패키지 구조 리팩터링
