+++ backend/src/main/java/com/bether/bether/connection/discord/application/service/DiscordSlashCommandRegistrar.java
@@ -1,4 +1,4 @@
-package com.bether.bether.discord.application.service;
+package com.bether.bether.connection.discord.application.service;
 
 import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.entities.Guild;
