+++ backend/src/main/java/com/bether/bether/connection/slack/application/dto/SlackRoomCreatedInput.java
@@ -1,4 +1,4 @@
-package com.bether.bether.slack.application.service.dto;
+package com.bether.bether.connection.slack.application.dto;
 
 import com.bether.bether.room.presentation.dto.request.RoomCreateRequest;
 import com.bether.bether.room.presentation.dto.response.RoomCreateResponse;
